Update calculate_coverage.sh
Since most companies use the cDNA synthesis sequencing technique, which is different from that of Vertis, I edited the calculated_coverage.sh script to adjust the direction of the strands for visualization in JBrowse. I also kept the original calculated_coverage.sh by renaming it to calculate_coverage_vertis.sh.